% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_heatmap.R
\name{pathway_heatmap}
\alias{pathway_heatmap}
\title{Creates a pretty heatmap with q-values for multiple GSEA comparisons
 expects a matrix containing normalized enrichment scores and q-values from gsea_to_mtx()}
\usage{
pathway_heatmap(nes, q = 0.1, clustrow = 1, clustcol = 1,
  cluster_cols = TRUE, main = NULL, annotlegend, samples = NULL,
  fontsize_row = 12, fontsize_col = 12)
}
\arguments{
\item{nes}{matrix of GSEA normalized enrichment scores and q-values created with gsea_to_mtx()}

\item{q}{minimum FDR / q-value cutoff for including genesets in the output matrix}

\item{clustrow}{number of clusters for the row dendogram (see pheatmap)}

\item{clustcol}{number of clusters for the row dendogram (see pheatmap)}

\item{main}{text string for main heatmap title}

\item{annotlegend}{boolean to include the annotation_legend or not (see pheatmap)}

\item{samples}{sample names to be used in the annotation_col, should be in the same order as the data matrix (see pheatmap)}

\item{fontsize_row}{font size for row text displaying gene sets (see pheatmap)}

\item{fontsize_col}{font size for col text displaying sample names (see pheatmap)}
}
\value{
a heatmap object for printing
}
\description{
Creates a heatmap (using pheatmap) from the matrix, with a primary heatmap of GSEA enrichment scores
  and a secondary heatmap of q-values (FDR). 
  A more stringent q-value can be set then used in gsea_to_mtx for subsequent heatmaps
}
\examples{
alldirs = list.dirs(path=dir_gsea3, full.names=TRUE, recursive=FALSE)
nes <- gsea_to_mtx(dirlist, db="GO", q=0.1, preN=50, postN=50)
pathway_heatmap(nes, clustrow=clustrow, clustcol=1, cluster_cols=FALSE,
                main=main, samples=samples, annotlegend=FALSE, 
                fontsize_row=fontsize_row, fontsize_col=12)

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{GSEA}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{geneset}
\keyword{heatmap}
